HOGPサーバのOn/Off設定をSharedPreferenceに保存するように修正 クラス名を変更
