+++ dConnectDevicePlugin/dConnectDeviceHOGP/plugin/src/main/java/org/deviceconnect/android/deviceplugin/hogp/server/HOGPServer.java
@@ -1,934 +1,188 @@
 package org.deviceconnect.android.deviceplugin.hogp.server;
 
-import android.annotation.TargetApi;
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.bluetooth.BluetoothGatt;
-import android.bluetooth.BluetoothGattCharacteristic;
-import android.bluetooth.BluetoothGattDescriptor;
-import android.bluetooth.BluetoothGattServer;
-import android.bluetooth.BluetoothGattServerCallback;
-import android.bluetooth.BluetoothGattService;
-import android.bluetooth.BluetoothManager;
-import android.bluetooth.BluetoothProfile;
-import android.bluetooth.le.AdvertiseCallback;
-import android.bluetooth.le.AdvertiseData;
-import android.bluetooth.le.AdvertiseSettings;
-import android.bluetooth.le.BluetoothLeAdvertiser;
-import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.os.Build;
-import android.os.Handler;
-import android.os.HandlerThread;
-import android.os.ParcelUuid;
-import android.support.v4.BuildConfig;
-import android.util.Log;
 
-import org.deviceconnect.android.deviceplugin.hogp.util.BatteryUtils;
-import org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.COLLECTION;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.END_COLLECTION;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.INPUT;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.LOGICAL_MAXIMUM;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.LOGICAL_MINIMUM;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.OUTPUT;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.REPORT_COUNT;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.REPORT_SIZE;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.USAGE;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.USAGE_MAXIMUM;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.USAGE_MINIMUM;
+import static org.deviceconnect.android.deviceplugin.hogp.util.HIDUtils.USAGE_PAGE;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Queue;
-import java.util.Set;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentLinkedQueue;
-
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_BATTERY_LEVEL;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_HID_CONTROL_POINT;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_HID_INFORMATION;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_MANUFACTURER_NAME;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_MODEL_NUMBER;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_PROTOCOL_MODE;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_REPORT;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_REPORT_MAP;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.CHARACTERISTIC_SERIAL_NUMBER;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.DESCRIPTOR_CLIENT_CHARACTERISTIC_CONFIGURATION;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.DESCRIPTOR_REPORT_REFERENCE;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.DEVICE_INFO_MAX_LENGTH;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.SERVICE_BATTERY;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.SERVICE_BLE_HID;
-import static org.deviceconnect.android.deviceplugin.hogp.util.BleUuidUtils.SERVICE_DEVICE_INFORMATION;
-
-@TargetApi(Build.VERSION_CODES.LOLLIPOP)
-public abstract class HOGPServer {
+public class HOGPServer extends AbstractHOGPServer {
     /**
-     * デバッグ用フラグ.
+     * Characteristic Data(Report Map)
      */
-    private static final boolean DEBUG = BuildConfig.DEBUG;
+    private static final byte[] REPORT_MAP = {
+            USAGE_PAGE(1),      0x01,         // Generic Desktop
+            USAGE(1),           0x02,         // Mouse
+            COLLECTION(1),      0x01,         // Application
+            USAGE(1),           0x01,         //  Pointer
+            COLLECTION(1),      0x00,         //  Physical
+            USAGE_PAGE(1),      0x09,         //   Buttons
+            USAGE_MINIMUM(1),   0x01,
+            USAGE_MAXIMUM(1),   0x03,
+            LOGICAL_MINIMUM(1), 0x00,
+            LOGICAL_MAXIMUM(1), 0x01,
+            REPORT_COUNT(1),    0x03,         //   3 bits (Buttons)
+            REPORT_SIZE(1),     0x01,
+            INPUT(1),           0x02,         //   Data, Variable, Absolute
 
-    /**
-     * デバッグ用タグ.
-     */
-    private static final String TAG = "HOGP";
+            REPORT_COUNT(1),    0x01,         //   5 bits (Padding)
+            REPORT_SIZE(1),     0x05,
+            INPUT(1),           0x01,         //   Constant
 
-    /**
-     * このインスタンスが属するコンテキスト.
-     */
-    private final Context mApplicationContext;
+            USAGE_PAGE(1),      0x01,         //   Generic Desktop
+            USAGE(1),           0x30,         //   X
+            USAGE(1),           0x31,         //   Y
+            USAGE(1),           0x38,         //   Wheel
+            LOGICAL_MINIMUM(1), (byte) 0x81,  //   -127
+            LOGICAL_MAXIMUM(1), 0x7f,         //   127
+            REPORT_SIZE(1),     0x08,         //   Three bytes
+            REPORT_COUNT(1),    0x03,
+            INPUT(1),           0x06,         //   Data, Variable, Relative
+            END_COLLECTION(0),
+            END_COLLECTION(0),
 
-    /**
-     * アドバタイズを行うクラス.
-     */
-    private BluetoothLeAdvertiser mBluetoothLeAdvertiser;
 
-    /**
-     *
-     */
-    private BluetoothGattCharacteristic mInputReportCharacteristic;
+            COLLECTION(1),      0x01,       // Application
+            USAGE_PAGE(1),      0x07,       //   keyboard/Keypad
+            USAGE_MINIMUM(1), (byte) 0xE0,
+            USAGE_MAXIMUM(1), (byte) 0xE7,
+            LOGICAL_MINIMUM(1), 0x00,
+            LOGICAL_MAXIMUM(1), 0x01,
+            REPORT_SIZE(1),     0x01,       //   1 byte (Modifier)
+            REPORT_COUNT(1),    0x08,
+            INPUT(1),           0x02,       //   Data,Var,Abs,No Wrap,Linear,Preferred State,No Null Position
 
-    /**
-     * GattServerのインスタンス.
-     */
-    private BluetoothGattServer mGattServer;
+            REPORT_COUNT(1),    0x01,       //   1 byte (Reserved)
+            REPORT_SIZE(1),     0x08,
+            INPUT(1),           0x01,       //   Const,Array,Abs,No Wrap,Linear,Preferred State,No Null Position
 
-    /**
-     * 接続したBluetoothデバイス.
-     */
-    private final Map<String, BluetoothDevice> mBluetoothDevicesMap = new HashMap<>();
+            REPORT_COUNT(1),    0x05,       //   5 bits (Num lock, Caps lock, Scroll lock, Compose, Kana)
+            REPORT_SIZE(1),     0x01,
+            USAGE_PAGE(1),      0x08,       //   LEDs
+            USAGE_MINIMUM(1),   0x01,       //   Num Lock
+            USAGE_MAXIMUM(1),   0x05,       //   Kana
+            OUTPUT(1),          0x02,       //   Data,Var,Abs,No Wrap,Linear,Preferred State,No Null Position,Non-volatile
 
-    /**
-     *
-     */
-    private byte[] mManufacturerName;
+            REPORT_COUNT(1),    0x01,       //   3 bits (Padding)
+            REPORT_SIZE(1),     0x03,
+            OUTPUT(1),          0x01,       //   Const,Array,Abs,No Wrap,Linear,Preferred State,No Null Position,Non-volatile
 
-    /**
-     *
-     */
-    private byte[] mDeviceName;
+            REPORT_COUNT(1),    0x06,       //   6 bytes (Keys)
+            REPORT_SIZE(1),     0x08,
+            LOGICAL_MINIMUM(1), 0x00,
+            LOGICAL_MAXIMUM(1), 0x65,       //   101 keys
+            USAGE_PAGE(1),      0x07,       //   keyboard/Keypad
+            USAGE_MINIMUM(1),   0x00,
+            USAGE_MAXIMUM(1),   0x65,
+            INPUT(1),           0x00,       //   Data,Array,Abs,No Wrap,Linear,Preferred State,No Null Position
 
-    /**
-     *
-     */
-    private byte[] mSerialNumber;
+            END_COLLECTION(0),
+    };
 
-    /**
-     * HID Input Report
-     */
-    private final Queue<byte[]> mInputReportQueue = new ConcurrentLinkedQueue<>();
+    private static final byte[] EMPTY_REPORT = new byte[12];
 
-    /**
-     * Bluetoothデバイスの処理を同期的に処理するためのハンドラ.
-     */
-    private Handler mHandler;
+    private static final int KEY_PACKET_MODIFIER_KEY_INDEX = 4;
+    private static final int KEY_PACKET_KEY_INDEX = 6;
 
-    /**
-     * ハンドラ用のスレッド.
-     */
-    private HandlerThread mHandlerThread;
+    private byte[] mKeyboardReport = new byte[12];
+    private byte[] mMouseReport = new byte[12];
+    private byte[] mLastMouseReport = new byte[12];
 
-    /**
-     * デバイスに送信するタイマー.
-     */
-    private Timer mTimer;
+    public HOGPServer(final Context context) {
+        super(context);
+    }
 
-    /**
-     * デバイスにHIDを送る送信レート.
-     */
-    private int mDataSendingRate = 10;
+    @Override
+    byte[] getReportMap() {
+        return REPORT_MAP;
+    }
 
-    /**
-     * デバイス接続イベントを通知するリスナー.
-     */
-    private OnHOGPServerListener mOnHOGPServerListener;
 
-    /**
-     * 空のバイト配列.
-     */
-    private static final byte[] EMPTY_BYTES = {};
+    @Override
+    void onOutputReport(final byte[] outputReport) {
+    }
 
-    /**
-     * HIDインフォメーションのレスポンス.
-     */
-    private static final byte[] RESPONSE_HID_INFORMATION = {0x11, 0x01, 0x00, 0x03};
-
-    /**
-     * コンストラクタ.
-     * @param context このインスタンスが属するコンテキスト
-     * @throws UnsupportedOperationException BluetoothやBLEがサポートされていない場合に発生
-     */
-    public HOGPServer(final Context context) throws UnsupportedOperationException {
-        mApplicationContext = context.getApplicationContext();
-
-        final BluetoothManager btMgr = (BluetoothManager) mApplicationContext.getSystemService(Context.BLUETOOTH_SERVICE);
-        final BluetoothAdapter btAdapter = btMgr.getAdapter();
-        if (btAdapter == null) {
-            throw new UnsupportedOperationException("Bluetooth is not available.");
-        }
-
-        if (!btAdapter.isEnabled()) {
-            throw new UnsupportedOperationException("Bluetooth is disabled.");
-        }
-
-        if (!btAdapter.isMultipleAdvertisementSupported()) {
-            throw new UnsupportedOperationException("Bluetooth LE Advertising not supported on this device.");
-        }
-
-        mBluetoothLeAdvertiser = btAdapter.getBluetoothLeAdvertiser();
-        if (mBluetoothLeAdvertiser == null) {
-            throw new UnsupportedOperationException("Bluetooth LE Advertising not supported on this device.");
-        }
-
-        PackageManager pm = context.getPackageManager();
-        String appName = "";
-        int versionCode = 0;
-        try {
-            PackageInfo packageInfo = pm.getPackageInfo(context.getPackageName(), 0);
-            appName = packageInfo.applicationInfo.loadLabel(pm).toString();
-            versionCode = packageInfo.versionCode;
-        } catch (PackageManager.NameNotFoundException e) {
-            if (DEBUG) {
-               Log.w(TAG, "", e);
+    private boolean isZero(final byte[] array) {
+        for (int i = 0; i < array.length; i ++) {
+            if (array[i] != 0) {
+                return false;
             }
         }
-        setManufacturerName(mApplicationContext.getPackageName());
-        setDeviceName(appName);
-        setSerialNumber("" + versionCode);
+        return true;
     }
 
     /**
-     * HIDデータ送信レートを設定します.
-     * @param dataSendingRate 送信レート
-     */
-    public void setDataSendingRate(final int dataSendingRate) {
-        mDataSendingRate = dataSendingRate;
-    }
-
-    /**
-     * OnHOGPServerListenerを設定します.
-     * @param listener リスナー
-     */
-    public void setOnHOGPServerListener(final OnHOGPServerListener listener) {
-        mOnHOGPServerListener = listener;
-    }
-
-    /**
-     * Manufacturer名を設定します.
+     * Mouseのレポートを送信します.
      *
-     * @param name Manufacturer名
+     * @param dx X座標の移動量 (-127 〜 127)
+     * @param dy Y座標の移動量 (-127 〜 127)
+     * @param wheel wheelの移動量 (-127 〜 127)
+     * @param leftButton 左ボタンが押下されている場合はtrue
+     * @param rightButton 右ボタンが押下されている場合はtrue
+     * @param middleButton 真ん中ボタンが押下されている場合はtrue
      */
-    public void setManufacturerName(final String name) {
-        mSerialNumber = convertString2Bytes(name);
-    }
+    public void movePointer(int dx, int dy, int wheel, final boolean leftButton, final boolean rightButton, final boolean middleButton) {
+        if (dx > 127) dx = 127;
+        if (dx < -127) dx = -127;
+        if (dy > 127) dy = 127;
+        if (dy < -127) dy = -127;
+        if (wheel > 127) wheel = 127;
+        if (wheel < -127) wheel = -127;
 
-    /**
-     * デバイス名を設定します.
-     *
-     * @param name the name
-     */
-    public void setDeviceName(final String name) {
-        mSerialNumber = convertString2Bytes(name);
-    }
-
-    /**
-     * シリアル番号を設定します.
-     *
-     * @param serialNumber the number
-     */
-    public final void setSerialNumber(final String serialNumber) {
-        mSerialNumber = convertString2Bytes(serialNumber);
-    }
-
-    /**
-     * 文字列をバイト配列に変換します.
-     * @param name 変換する文字列
-     * @return 変換されたバイト配列
-     */
-    private byte[] convertString2Bytes(String name) {
-        final byte[] nameBytes = name.getBytes(StandardCharsets.UTF_8);
-        if (nameBytes.length > DEVICE_INFO_MAX_LENGTH) {
-            final byte[] bytes = new byte[DEVICE_INFO_MAX_LENGTH];
-            System.arraycopy(nameBytes, 0, bytes, 0, DEVICE_INFO_MAX_LENGTH);
-            return bytes;
-        } else {
-            return nameBytes;
+        byte button = 0;
+        if (leftButton) {
+            button |= 1;
         }
-    }
-
-    /**
-     * Represents Report Map byte array
-     * @return Report Map data
-     */
-    abstract byte[] getReportMap();
-
-    /**
-     * HID Output Report
-     *
-     * @param outputReport the report data
-     */
-    abstract void onOutputReport(final byte[] outputReport);
-
-    /**
-     * 送信するHIDデータを追加します.
-     * @param inputReport 追加するデータ
-     */
-    final void addInputReport(final byte[] inputReport) {
-        if (inputReport != null && inputReport.length > 0) {
-            mInputReportQueue.offer(inputReport);
+        if (rightButton) {
+            button |= 2;
         }
-    }
+        if (middleButton) {
+            button |= 4;
+        }
 
-    /**
-     * HOGPサーバを開始します.
-     */
-    public void start() {
+        mMouseReport[0] = (byte) (button & 7);
+        mMouseReport[1] = (byte) dx;
+        mMouseReport[2] = (byte) dy;
+        mMouseReport[3] = (byte) wheel;
 
-        if (mGattServer != null) {
+        if (isZero(mLastMouseReport) && isZero(mMouseReport)) {
             return;
         }
 
-        final BluetoothManager btMgr = (BluetoothManager) mApplicationContext.getSystemService(Context.BLUETOOTH_SERVICE);
-        mGattServer = btMgr.openGattServer(mApplicationContext, mGattServerCallback);
-        if (mGattServer == null) {
-            throw new UnsupportedOperationException("mGattServer is null, check Bluetooth is ON.");
-        }
+        addInputReport(mMouseReport);
 
-        mHandlerThread = new HandlerThread(TAG);
-        mHandlerThread.start();
-        mHandler = new Handler(mHandlerThread.getLooper());
-
-        addService(setUpHidService(true, true, false));
-        addService(setUpDeviceInformationService());
-//        addService(setUpBatteryService());
-
-        mTimer = new Timer();
-        mTimer.scheduleAtFixedRate(new TimerTask() {
-            @Override
-            public void run() {
-                final byte[] polled = mInputReportQueue.poll();
-                if (polled != null && mInputReportCharacteristic != null) {
-                    mInputReportCharacteristic.setValue(polled);
-                    mHandler.post(new Runnable() {
-                        @Override
-                        public void run() {
-                            final Set<BluetoothDevice> devices = getDevices();
-                            for (final BluetoothDevice device : devices) {
-                                try {
-                                    if (mGattServer != null) {
-                                        mGattServer.notifyCharacteristicChanged(device, mInputReportCharacteristic, false);
-                                    }
-                                } catch (final Throwable ignored) {
-                                    // do nothing.
-                                }
-                            }
-                        }
-                    });
-                }
-            }
-        }, 0, mDataSendingRate);
-
-        startAdvertising();
+        mLastMouseReport[0] = mMouseReport[0];
+        mLastMouseReport[1] = mMouseReport[1];
+        mLastMouseReport[2] = mMouseReport[2];
+        mLastMouseReport[3] = mMouseReport[3];
     }
 
     /**
-     * HOGPサーバを停止します.
-     */
-    public void stop() {
-        stopAdvertising();
-    }
-
-    /**
-     * BluetoothGattServerにサービスを追加します.
+     * キーダウンのレポートを送信します.
      *
-     * @param service 追加するサービス
+     * @param modifier modifier key
+     * @param keyCode key code
      */
-    private void addService(final BluetoothGattService service) {
-        boolean serviceAdded = false;
-        while (!serviceAdded) {
-            try {
-                serviceAdded = mGattServer.addService(service);
-            } catch (final Exception e) {
-                if (DEBUG) {
-                    Log.w(TAG, "Adding Service failed", e);
-                }
-            }
-        }
-        if (DEBUG) {
-            Log.d(TAG, "Service: " + service.getUuid() + " added.");
-        }
+    public void sendKeyDown(final byte modifier, final byte keyCode) {
+        mKeyboardReport[KEY_PACKET_MODIFIER_KEY_INDEX] = modifier;
+        mKeyboardReport[KEY_PACKET_KEY_INDEX] = keyCode;
+        addInputReport(mKeyboardReport);
     }
 
     /**
-     * Device Informationサービスを設定します.
-     *
-     * @return Device Information サービスを
+     * キーあっぷのレポートを送信します.
      */
-    private BluetoothGattService setUpDeviceInformationService() {
-        final BluetoothGattService service = new BluetoothGattService(SERVICE_DEVICE_INFORMATION, BluetoothGattService.SERVICE_TYPE_PRIMARY);
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_MANUFACTURER_NAME,
-                    BluetoothGattCharacteristic.PROPERTY_READ,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED);
-            while (!service.addCharacteristic(characteristic));
-        }
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_MODEL_NUMBER,
-                    BluetoothGattCharacteristic.PROPERTY_READ,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED);
-            while (!service.addCharacteristic(characteristic));
-        }
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_SERIAL_NUMBER,
-                    BluetoothGattCharacteristic.PROPERTY_READ,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED);
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        return service;
+    public void sendKeyUp() {
+        addInputReport(EMPTY_REPORT);
     }
 
-    /**
-     * Setup Battery Service
-     *
-     * @return the service
-     */
-    private BluetoothGattService setUpBatteryService() {
-        final BluetoothGattService service = new BluetoothGattService(SERVICE_BATTERY, BluetoothGattService.SERVICE_TYPE_PRIMARY);
-
-        // Battery Level
-        final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                CHARACTERISTIC_BATTERY_LEVEL,
-                BluetoothGattCharacteristic.PROPERTY_NOTIFY | BluetoothGattCharacteristic.PROPERTY_READ,
-                BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED);
-
-        final BluetoothGattDescriptor clientCharacteristicConfigurationDescriptor = new BluetoothGattDescriptor(
-                DESCRIPTOR_CLIENT_CHARACTERISTIC_CONFIGURATION,
-                BluetoothGattDescriptor.PERMISSION_READ | BluetoothGattDescriptor.PERMISSION_WRITE);
-        clientCharacteristicConfigurationDescriptor.setValue(BluetoothGattDescriptor.ENABLE_NOTIFICATION_VALUE);
-        characteristic.addDescriptor(clientCharacteristicConfigurationDescriptor);
-
-        while (!service.addCharacteristic(characteristic));
-
-        return service;
-    }
-
-    /**
-     * HIDサービスを設定します.
-     *
-     * @param isNeedInputReport true: serves 'Input Report' BLE characteristic
-     * @param isNeedOutputReport true: serves 'Output Report' BLE characteristic
-     * @param isNeedFeatureReport true: serves 'Feature Report' BLE characteristic
-     * @return HIDサービス
-     */
-    private BluetoothGattService setUpHidService(final boolean isNeedInputReport, final boolean isNeedOutputReport, final boolean isNeedFeatureReport) {
-        final BluetoothGattService service = new BluetoothGattService(SERVICE_BLE_HID, BluetoothGattService.SERVICE_TYPE_PRIMARY);
-
-        // HID Information
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_HID_INFORMATION,
-                    BluetoothGattCharacteristic.PROPERTY_READ,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED);
-
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        // Report Map
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_REPORT_MAP,
-                    BluetoothGattCharacteristic.PROPERTY_READ,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED);
-
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        // Protocol Mode
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_PROTOCOL_MODE,
-                    BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED | BluetoothGattCharacteristic.PERMISSION_WRITE_ENCRYPTED);
-            characteristic.setWriteType(BluetoothGattCharacteristic.WRITE_TYPE_NO_RESPONSE);
-
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        // HID Control Point
-        {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_HID_CONTROL_POINT,
-                    BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE,
-                    BluetoothGattCharacteristic.PERMISSION_WRITE_ENCRYPTED);
-            characteristic.setWriteType(BluetoothGattCharacteristic.WRITE_TYPE_NO_RESPONSE);
-
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        // Input Report
-        if (isNeedInputReport) {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_REPORT,
-                    BluetoothGattCharacteristic.PROPERTY_NOTIFY | BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED | BluetoothGattCharacteristic.PERMISSION_WRITE_ENCRYPTED);
-
-            final BluetoothGattDescriptor clientCharacteristicConfigurationDescriptor = new BluetoothGattDescriptor(
-                    DESCRIPTOR_CLIENT_CHARACTERISTIC_CONFIGURATION,
-                    BluetoothGattDescriptor.PERMISSION_READ_ENCRYPTED | BluetoothGattDescriptor.PERMISSION_WRITE_ENCRYPTED); //  | BluetoothGattDescriptor.PERMISSION_WRITE
-            clientCharacteristicConfigurationDescriptor.setValue(BluetoothGattDescriptor.ENABLE_NOTIFICATION_VALUE);
-            characteristic.addDescriptor(clientCharacteristicConfigurationDescriptor);
-
-            final BluetoothGattDescriptor reportReferenceDescriptor = new BluetoothGattDescriptor(
-                    DESCRIPTOR_REPORT_REFERENCE,
-                    BluetoothGattDescriptor.PERMISSION_READ_ENCRYPTED | BluetoothGattDescriptor.PERMISSION_WRITE_ENCRYPTED);
-            characteristic.addDescriptor(reportReferenceDescriptor);
-
-            while (!service.addCharacteristic(characteristic));
-            mInputReportCharacteristic = characteristic;
-        }
-
-        // Output Report
-        if (isNeedOutputReport) {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_REPORT,
-                    BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE | BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED | BluetoothGattCharacteristic.PERMISSION_WRITE_ENCRYPTED);
-            characteristic.setWriteType(BluetoothGattCharacteristic.WRITE_TYPE_NO_RESPONSE);
-
-            final BluetoothGattDescriptor descriptor = new BluetoothGattDescriptor(
-                    DESCRIPTOR_REPORT_REFERENCE,
-                    BluetoothGattDescriptor.PERMISSION_READ_ENCRYPTED | BluetoothGattDescriptor.PERMISSION_WRITE_ENCRYPTED);
-            characteristic.addDescriptor(descriptor);
-
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        // Feature Report
-        if (isNeedFeatureReport) {
-            final BluetoothGattCharacteristic characteristic = new BluetoothGattCharacteristic(
-                    CHARACTERISTIC_REPORT,
-                    BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE,
-                    BluetoothGattCharacteristic.PERMISSION_READ_ENCRYPTED | BluetoothGattCharacteristic.PERMISSION_WRITE_ENCRYPTED);
-
-            final BluetoothGattDescriptor descriptor = new BluetoothGattDescriptor(
-                    DESCRIPTOR_REPORT_REFERENCE,
-                    BluetoothGattDescriptor.PERMISSION_READ_ENCRYPTED | BluetoothGattDescriptor.PERMISSION_WRITE_ENCRYPTED);
-            characteristic.addDescriptor(descriptor);
-
-            while (!service.addCharacteristic(characteristic));
-        }
-
-        return service;
-    }
-
-    /**
-     * アドバタイジングを開始します.
-     */
-    private void startAdvertising() {
-        if (DEBUG) {
-            Log.d(TAG, "startAdvertising");
-        }
-
-        mHandler.post(new Runnable() {
-            @Override
-            public void run() {
-                // set up advertising setting
-                final AdvertiseSettings advertiseSettings = new AdvertiseSettings.Builder()
-                        .setTxPowerLevel(AdvertiseSettings.ADVERTISE_TX_POWER_HIGH)
-                        .setConnectable(true)
-                        .setTimeout(0)
-                        .setAdvertiseMode(AdvertiseSettings.ADVERTISE_MODE_LOW_LATENCY)
-                        .build();
-
-                // set up advertising data
-                final AdvertiseData advertiseData = new AdvertiseData.Builder()
-                        .setIncludeTxPowerLevel(false)
-                        .setIncludeDeviceName(true)
-                        .addServiceUuid(ParcelUuid.fromString(SERVICE_DEVICE_INFORMATION.toString()))
-                        .addServiceUuid(ParcelUuid.fromString(SERVICE_BLE_HID.toString()))
-                        .addServiceUuid(ParcelUuid.fromString(SERVICE_BATTERY.toString()))
-                        .build();
-
-                // set up scan result
-                final AdvertiseData scanResult = new AdvertiseData.Builder()
-                        .addServiceUuid(ParcelUuid.fromString(SERVICE_DEVICE_INFORMATION.toString()))
-                        .addServiceUuid(ParcelUuid.fromString(SERVICE_BLE_HID.toString()))
-                        .addServiceUuid(ParcelUuid.fromString(SERVICE_BATTERY.toString()))
-                        .build();
-
-                mBluetoothLeAdvertiser.startAdvertising(advertiseSettings, advertiseData, scanResult, mAdvertiseCallback);
-            }
-        });
-    }
-
-    /**
-     * アドバタイジングを停止します.
-     */
-    private void stopAdvertising() {
-        if (DEBUG) {
-            Log.d(TAG, "stopAdvertising");
-        }
-
-        mHandler.post(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    mBluetoothLeAdvertiser.stopAdvertising(mAdvertiseCallback);
-                } catch (final IllegalStateException ignored) {
-                    // BT Adapter is not turned ON
-                }
-
-                try {
-                    if (mTimer != null) {
-                        mTimer.cancel();
-                        mTimer = null;
-                    }
-
-                    if (mGattServer != null) {
-                        final Collection<BluetoothDevice> devices = getDevices();
-                        for (final BluetoothDevice device : devices) {
-                            mGattServer.cancelConnection(device);
-                        }
-
-                        mGattServer.clearServices();
-                        mGattServer.close();
-                        mGattServer = null;
-                    }
-                } catch (final IllegalStateException ignored) {
-                    // do nothing
-                }
-
-                mHandlerThread.quit();
-                mHandlerThread = null;
-            }
-        });
-    }
-
-    /**
-     * Obtains connected Bluetooth devices
-     *
-     * @return the connected Bluetooth devices
-     */
-    private Set<BluetoothDevice> getDevices() {
-        final Set<BluetoothDevice> deviceSet = new HashSet<>();
-        synchronized (mBluetoothDevicesMap) {
-            deviceSet.addAll(mBluetoothDevicesMap.values());
-        }
-        return Collections.unmodifiableSet(deviceSet);
-    }
-
-    /**
-     * 指定されてBluetoothデバイスとペアリングを行います.
-     * @param device ペアリングを行うBluetoothデバイス
-     */
-    private void paringDevice(final BluetoothDevice device) {
-        if (DEBUG) {
-            Log.i(TAG, "AAAAAAAAAAAAAAAAAAAA paringDevice.");
-        }
-
-        mApplicationContext.registerReceiver(new BroadcastReceiver() {
-            @Override
-            public void onReceive(final Context context, final Intent intent) {
-                final String action = intent.getAction();
-                if (BluetoothDevice.ACTION_BOND_STATE_CHANGED.equals(action)) {
-                    int state = intent.getIntExtra(BluetoothDevice.EXTRA_BOND_STATE, BluetoothDevice.ERROR);
-                    switch (state) {
-                        case BluetoothDevice.BOND_BONDING:
-                            if (DEBUG) {
-                                Log.e("ABC", "AAAAAAAAAAAAAAAAAAAA bond bonding.");
-                            }
-                            break;
-
-                        case BluetoothDevice.BOND_BONDED:
-                            BluetoothDevice bondedDevice = intent.getParcelableExtra(BluetoothDevice.EXTRA_DEVICE);
-                            if (!bondedDevice.getAddress().equals(device.getAddress())) {
-                                // 違うデバイスと接続された場合
-                                if (DEBUG) {
-                                    Log.w("ABC", "BBBBBBBBBBBBBBBBBBB Connected to a different device");
-                                }
-                                return;
-                            }
-
-                            context.unregisterReceiver(this);
-                            connectDevice(device);
-                            break;
-
-                        default:
-                            if (DEBUG) {
-                                Log.e(TAG, "AAAAAAAAAAAAAAAAAAAA bond error.");
-                            }
-                            context.unregisterReceiver(this);
-                            break;
-                    }
-                }
-            }
-        }, new IntentFilter(BluetoothDevice.ACTION_BOND_STATE_CHANGED));
-
-        try {
-            device.setPairingConfirmation(true);
-        } catch (Exception e) {
-            if (DEBUG) {
-                Log.d(TAG, e.getMessage(), e);
-            }
-        }
-
-        device.createBond();
-    }
-
-    /**
-     * 指定されたBluetoothデバイスと接続を行います.
-     * @param device 接続を行うBluetoothデバイス
-     */
-    private void connectDevice(final BluetoothDevice device) {
-        mHandler.post(new Runnable() {
-            @Override
-            public void run() {
-                if (mGattServer != null) {
-                    mGattServer.connect(device, true);
-                }
-            }
-        });
-
-        synchronized (mBluetoothDevicesMap) {
-            mBluetoothDevicesMap.put(device.getAddress(), device);
-        }
-
-        if (mOnHOGPServerListener != null) {
-            mOnHOGPServerListener.onConnected(device);
-        }
-    }
-
-    /**
-     * 指定されたBluetoothデバイスが切断された時の処理を行います.
-     * <p>
-     * 切断されたデバイスは、再接続処理を行います。
-     * </p>
-     * @param device 切断されたデバイス
-     */
-    private void disconnectDevice(final BluetoothDevice device) {
-        mHandler.post(new Runnable() {
-            @Override
-            public void run() {
-                if (mGattServer != null) {
-                    mGattServer.connect(device, true);
-                }
-            }
-        });
-
-        synchronized (mBluetoothDevicesMap) {
-            mBluetoothDevicesMap.remove(device.getAddress());
-        }
-
-        if (mOnHOGPServerListener != null) {
-            mOnHOGPServerListener.onDisconnected(device);
-        }
-    }
-
-    /**
-     * アドバタイジング用のコールバック.
-     * <p>
-     * 特に何も処理しません。
-     * </p>
-     */
-    private final AdvertiseCallback mAdvertiseCallback = new AdvertiseCallback() {};
-
-    /**
-     * BluetoothGattServerからの通知を受け取るコールバック.
-     */
-    private final BluetoothGattServerCallback mGattServerCallback = new BluetoothGattServerCallback() {
-
-        @Override
-        public void onConnectionStateChange(final BluetoothDevice device, final int status, final int newState) {
-            if (DEBUG) {
-                Log.d(TAG, "onConnectionStateChange status: " + status + ", newState: " + newState);
-            }
-
-            if (mGattServer == null) {
-                return;
-            }
-
-            switch (newState) {
-                case BluetoothProfile.STATE_CONNECTED:
-                    if (DEBUG) {
-                        Log.d(TAG, "BluetoothProfile.STATE_CONNECTED bondState: " + device.getBondState());
-                    }
-
-                    // check bond status
-                    if (device.getBondState() == BluetoothDevice.BOND_NONE) {
-                        paringDevice(device);
-                    } else if (device.getBondState() == BluetoothDevice.BOND_BONDED) {
-                        connectDevice(device);
-                    }
-                    break;
-
-                case BluetoothProfile.STATE_DISCONNECTED:
-                    if (DEBUG) {
-                        Log.w(TAG, "BluetoothProfile.STATE_DISCONNECTED");
-                    }
-                    disconnectDevice(device);
-                    break;
-
-                default:
-                    // do nothing
-                    break;
-            }
-        }
-
-        @Override
-        public void onCharacteristicReadRequest(final BluetoothDevice device, final int requestId, final int offset,
-                                                final BluetoothGattCharacteristic characteristic) {
-            if (DEBUG) {
-                Log.d(TAG, "onCharacteristicReadRequest characteristic: " + characteristic.getUuid() + ", offset: " + offset);
-            }
-
-            if (mGattServer == null) {
-                return;
-            }
-
-            mHandler.post(new Runnable() {
-                @Override
-                public void run() {
-                    final UUID characteristicUuid = characteristic.getUuid();
-                    if (BleUuidUtils.matches(CHARACTERISTIC_HID_INFORMATION, characteristicUuid)) {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, RESPONSE_HID_INFORMATION);
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_REPORT_MAP, characteristicUuid)) {
-                        if (offset == 0) {
-                            mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, getReportMap());
-                        } else {
-                            final int remainLength = getReportMap().length - offset;
-                            if (remainLength > 0) {
-                                final byte[] data = new byte[remainLength];
-                                System.arraycopy(getReportMap(), offset, data, 0, remainLength);
-                                mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, offset, data);
-                            } else {
-                                mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, offset, null);
-                            }
-                        }
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_HID_CONTROL_POINT, characteristicUuid)) {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, new byte []{0});
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_REPORT, characteristicUuid)) {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, EMPTY_BYTES);
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_MANUFACTURER_NAME, characteristicUuid)) {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, mManufacturerName);
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_SERIAL_NUMBER, characteristicUuid)) {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, mSerialNumber);
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_MODEL_NUMBER, characteristicUuid)) {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, mDeviceName);
-                    } else if (BleUuidUtils.matches(CHARACTERISTIC_BATTERY_LEVEL, characteristicUuid)) {
-                        byte level = (byte) (BatteryUtils.getBatteryLevel(mApplicationContext) * 100);
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, new byte[] {level});
-                    } else {
-                        mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, characteristic.getValue());
-                    }
-                }
-            });
-        }
-
-        @Override
-        public void onDescriptorReadRequest(final BluetoothDevice device, final int requestId, final int offset,
-                                            final BluetoothGattDescriptor descriptor) {
-            if (DEBUG) {
-                Log.d(TAG, "onDescriptorReadRequest requestId: " + requestId + ", offset: " + offset + ", descriptor: " + descriptor.getUuid());
-            }
-
-            if (mGattServer == null) {
-                return;
-            }
-
-            mHandler.post(new Runnable() {
-                @Override
-                public void run() {
-                    if (BleUuidUtils.matches(DESCRIPTOR_REPORT_REFERENCE, descriptor.getUuid())) {
-                        final int characteristicProperties = descriptor.getCharacteristic().getProperties();
-                        if (characteristicProperties == (BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE | BluetoothGattCharacteristic.PROPERTY_NOTIFY)) {
-                            mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, new byte[]{0, 1});
-                        } else if (characteristicProperties == (BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE | BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE)) {
-                            mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, new byte[]{0, 2});
-                        } else if (characteristicProperties == (BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE)) {
-                            mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, new byte[]{0, 3});
-                        } else {
-                            mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_FAILURE, 0, EMPTY_BYTES);
-                        }
-                    }
-                }
-            });
-        }
-
-        @Override
-        public void onCharacteristicWriteRequest(final BluetoothDevice device, final int requestId,
-                                                 final BluetoothGattCharacteristic characteristic,
-                                                 final boolean preparedWrite, final boolean responseNeeded,
-                                                 final int offset, final byte[] value) {
-            if (DEBUG) {
-                Log.d(TAG, "onCharacteristicWriteRequest characteristic: " + characteristic.getUuid() + ", value: " + Arrays.toString(value));
-            }
-
-            if (mGattServer == null) {
-                return;
-            }
-
-            if (responseNeeded) {
-                if (BleUuidUtils.matches(CHARACTERISTIC_REPORT, characteristic.getUuid())) {
-                    if (characteristic.getProperties() == (BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_WRITE | BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE)) {
-                        // Output Report
-                        onOutputReport(value);
-                    }
-                    mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, EMPTY_BYTES);
-                }
-            }
-        }
-
-        @Override
-        public void onDescriptorWriteRequest(final BluetoothDevice device, final int requestId,
-                                             final BluetoothGattDescriptor descriptor, final boolean preparedWrite,
-                                             final boolean responseNeeded, final int offset, final byte[] value) {
-            if (DEBUG) {
-                Log.d(TAG, "onDescriptorWriteRequest descriptor: " + descriptor.getUuid() + ", value: " + Arrays.toString(value) + ", responseNeeded: " + responseNeeded + ", preparedWrite: " + preparedWrite);
-            }
-
-            if (mGattServer == null) {
-                return;
-            }
-
-            descriptor.setValue(value);
-
-            if (responseNeeded) {
-                if (BleUuidUtils.matches(DESCRIPTOR_CLIENT_CHARACTERISTIC_CONFIGURATION, descriptor.getUuid())) {
-                    mGattServer.sendResponse(device, requestId, BluetoothGatt.GATT_SUCCESS, 0, EMPTY_BYTES);
-                }
-            }
-        }
-
-        @Override
-        public void onServiceAdded(final int status, final BluetoothGattService service) {
-            if (DEBUG) {
-                Log.d(TAG, "onServiceAdded status: " + status + ", service: " + service.getUuid());
-                if (status != 0) {
-                    Log.d(TAG, "onServiceAdded Adding Service failed..");
-                }
-            }
-        }
-    };
-
-
-    /**
-     * HOGPServerにデバイスが接続・切断した時の通知を行うリスナー.
-     */
-    public interface OnHOGPServerListener {
-        /**
-         * デバイスが接続された時に呼び出すメソッド.
-         * @param device 接続したデバイス
-         */
-        void onConnected(BluetoothDevice device);
-
-        /**
-         * デバイスが切断された時に呼び出すメソッド.
-         * @param device 切断したデバイス
-         */
-        void onDisconnected(BluetoothDevice device);
-    }
 }
